Componentize the tablature dash text area
